Relocate currency helpers to utils module
